BUR-168 resolve PR comments (2/2)
